WIP Start migrating UI code to use the new data model.
Still not compiling, we're 50% there
