+++ blackjack_ui/src/graph.rs
@@ -1,5 +1,2 @@
 /// Provides implementations to the traits in `egui_node_graph` specific to blackjack
 pub mod node_graph;
-
-/// Compiles node graphs into Lua
-pub mod graph_compiler;
